refactor: Flatten sections, move sections/ child components to sections/common/
