--- src/app/contact/page.tsx
@@ -1,10 +1,0 @@
-import Overview from "@/components/sections/contact/Overview";
-
-export default function Contact() {
-  return (
-    <div className="flex flex-1 flex-col font-sans">
-      <Overview />
-    </div>
-  );
-}
-
