--- src/app/education/page.tsx
@@ -1,10 +1,0 @@
-import Overview from "@/components/sections/education/Overview";
-
-export default function Education() {
-  return (
-    <div className="flex flex-1 flex-col font-sans">
-      <Overview />
-    </div>
-  );
-}
-
