+++ src/app/page.tsx
@@ -1,9 +1,11 @@
-import Hero from "@/components/sections/about/Hero";
+import Introduction from "@/components/sections/Introduction";
+import WorkTimeline from "@/components/sections/Work";
 
 export default function Home() {
   return (
     <div className="flex flex-1 flex-col font-sans">
-      <Hero />
+      <Introduction />
+      <WorkTimeline />
     </div>
   );
 }
