--- src/app/projects/page.tsx
@@ -1,10 +1,0 @@
-import Overview from "@/components/sections/projects/Overview";
-
-export default function Projects() {
-  return (
-    <div className="flex flex-1 flex-col font-sans">
-      <Overview />
-    </div>
-  );
-}
-
