--- src/app/work/page.tsx
@@ -1,10 +1,0 @@
-import Overview from "@/components/sections/work/Overview";
-
-export default function Work() {
-  return (
-    <div className="flex flex-1 flex-col font-sans">
-      <Overview />
-    </div>
-  );
-}
-
