+++ src/components/sections/common/WorkItem.tsx
