+++ src/lib/site.ts
@@ -1,10 +1,6 @@
 // Navigation links for the website
 export const navLinks = [
   { href: "/", label: "About" },
-  { href: "/work", label: "Work" },
-  { href: "/education", label: "Education" },
-  { href: "/projects", label: "Projects" },
-  { href: "/contact", label: "Contact" },
 ];
 
 // Profile information
